Reorganize mock_uss flight functionality
