+++ monitoring/mock_uss/f3548v21/README.md
@@ -1,0 +1,3 @@
+# mock_uss: ASTM F3548-21
+
+[ASTM F3548-21](http://astm.org/f3548-21.html) standardizes UTM interoperability between USSs to achieve strategic coordination and communicate constraints.  This folder enables [mock_uss](..) to comply with the Strategic Conflict Detection requirements from that standard.
