+++ monitoring/mock_uss/flight_planning/README.md
@@ -1,0 +1,3 @@
+# mock_uss: flight_planner
+
+This folder contains materials implementing [InterUSS's flight_planning automated testing interface](https://github.com/interuss/automated_testing_interfaces/tree/main/flight_planning) by [mock_uss](..).
