+++ monitoring/mock_uss/flights/README.md
@@ -1,0 +1,3 @@
+# mock_uss: flights
+
+This folder contains materials related to generic handling of user-requested flights by [mock_uss](..).
