+++ monitoring/mock_uss/scd_injection/README.md
@@ -1,0 +1,3 @@
+# mock_uss: scd_injection
+
+This folder contains material related to the deprecated [InterUSS scd automated testing interface](https://github.com/interuss/automated_testing_interfaces/tree/main/scd).
