+++ monitoring/mock_uss/scd_injection/routes.py
@@ -1,0 +1,6 @@
+from monitoring.mock_uss import webapp
+
+
+@webapp.route("/scdsc/status")
+def scdsc_status():
+    return "scd flight injection API ok"
