--- monitoring/mock_uss/scdsc/README.md
@@ -1,1 +1,0 @@
-[ASTM F3548-21](http://astm.org/f3548-21.html) standardizes UTM interoperability between USSs to achieve strategic coordination and communicate constraints.  When this `scdsc` [mock_uss](..) functionality is enabled, mock_uss will behave like an F3548-21 strategic coordinator that accepts user flight planning attempts via the [InterUSS scd automated testing interface](../../../interfaces/automated_testing/scd) and attempts to establish and communicate an operational intent for those flights according to ASTM F3548-21.
